Django-Middleware-0x03/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
Pillow>=9.0.0
django-filter>=23.2
drf-nested-routers>=0.93.4
